Desktop/pic-n-eat-v2/ml_model/requirements.txt
# ML Model Requirements
torch>=2.0.0
torchvision>=0.15.0
pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
opencv-python>=4.8.0
albumentations>=1.3.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Data handling
requests>=2.31.0
openpyxl>=3.1.0

# Model optimization
timm>=0.9.0
torchmetrics>=1.0.0

# Deployment
onnx>=1.14.0
onnxruntime>=1.15.0
